bot_vs_bot inherits start method
