+++ models/abstracts/game_type.py
@@ -4,6 +4,7 @@ from ..board import Board
 from ..handle_player_input import HandlePlayerInput
 from ..handle_turns import HandleTurns
 from ..game_display import GameDisplay
+from ..human_player import HumanPlayer
 
 class GameType(metaclass=ABCMeta):
 
@@ -21,6 +22,28 @@ class GameType(metaclass=ABCMeta):
         self.start()
         GameDisplay.show(self.board)
         GameDisplay.log('Game Over')
+
+    def start(self):
+       while not self.gameState.finished(self.board):
+           if(self.handleTurns.currentPlayerToken == self.playerOne.token):
+             self.handle_play(self.playerOne)
+           else:
+               self.handle_play(self.playerTwo)
+       self.end_game()
+
+
+    def handle_play(self, player):
+        if( isinstance(player, HumanPlayer) ):
+             try_spot = self.handlePlayerInput.get_player_spot(self.board.get_available_spots())
+             spot = copy.deepcopy(try_spot)
+             player.play(self.board, spot)
+             GameDisplay.log(f"Human {self.playerOne.token} has played in spot {spot}")
+        else:
+             spot = self.playerTwo.play(self.board)
+             GameDisplay.log(f"Bot {self.playerOne.token} has played in spot {spot}")
+        GameDisplay.show(self.board)
+        self.handleTurns.change()
+
 
     def set_tokens(self):
       while self.playerTwo.token == None:
