+++ models/game_types/human_vs_bot.py
@@ -13,7 +13,7 @@ class HumanVsBot(GameType):
   def start(self):
      while not self.gameState.finished(self.board):
          if(self.handleTurns.currentPlayerToken == self.playerOne.token):
-             spot = self.let_playerOne_play()
+             spot = self.let_humanPlayer_play(self.playerOne)
              GameDisplay.log(f"Human {self.playerOne.token} has played in spot {spot}")
              GameDisplay.show(self.board)
              self.handleTurns.change()
@@ -23,9 +23,9 @@ class HumanVsBot(GameType):
              GameDisplay.show(self.board)
              self.handleTurns.change()
          self.end_game()
-
-  def let_playerOne_play(self):
+         
+  def let_humanPlayer_play(self, player):
       try_spot = self.handlePlayerInput.get_player_spot(self.board.get_available_spots())
       spot = copy.deepcopy(try_spot)
-      self.playerOne.play(self.board, spot)
+      player.play(self.board, spot)
       return spot
